feat: role a11y fixes, documentation (#126)
BREAKING CHANGE:
The Grid component used to automatically add the appropriate role for a grid on each of its components (grid --> row --> gridcell). We are use the Grid component extensively, but just for layout purposes, which causes accessibility issues. Now, by default, the Grid component does not include any of these roles, so that it can also be used for layouts. You can add the roles back in by passing roles=true on the Grid component.